Split selection_rate into its own file
